refactor: split workflow integration tests into 14 topical modules
- Split 5286-line workflow.rs into workflow/{analysis,capabilities,change,
  dispatch,harness,impact,jobs,misc,onboard,resources,schema,session,symbol,
  workflow}.rs
- Move shared PATH_MUTEX and prepend_path into workflow/mod.rs
- All 562 tests pass, clippy clean
